+++ src/errors.ts
@@ -1,19 +1,27 @@
 export interface PochoirErrorOptions extends ErrorOptions {
     verbose?: boolean;
     notice?: boolean;
+    prefix?: string;
 }
 
 export class PochoirError extends Error {
     verbose: boolean;
     notice: boolean;
+    prefix: string;
 
     constructor(
         msg: string,
-        { verbose = true, notice = true, ...options }: PochoirErrorOptions = {},
+        {
+            verbose = true,
+            notice = true,
+            prefix = "Pochoir Error",
+            ...options
+        }: PochoirErrorOptions = {},
     ) {
         super(msg, options);
         this.verbose = verbose;
         this.notice = notice;
+        this.prefix = prefix;
     }
 }