+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- Add `$.options` property to control template creation behavior. `$.options` accepts the following list of strings:
+    - `confirmName` will prompt a modal to confirm note name before creation
+    - `openIfExists` will not render the template if note already exists
+
 ## [0.0.21] - 2026-02-03
 
 ### Added